PCG-биомы, шаг 4: взвешенный спавн ингредиентов на стыке биомов
GetRandomResourceForBiome теперь сливает кандидатов из всех долей
Cell.BiomeWeights (не одно точное совпадение по Cell.Biome): для каждой
доли (Biome, Weight) берёт кэш кандидатов+весов, домножает базовый вес
карточки на долю клетки, сливает всё в один список, один взвешенный
ролл через PickWeightedResource -- не отдельный RNG-бросок "какой биом
сначала", сохраняет однодетерминированную схему пайплайна. Пустой
Cell.BiomeWeights (клетка вне всех PCG-регионов) -- обязательный откат на
{Cell.Biome: 1.0}, не опция: сохраняет побитовую совместимость с 8
существующими тестами, которые конструируют FGridCell напрямую и
выставляют только Biome.

Правит реальный баг, найденный на этапе планирования (не в рантайме):
PickWeightedResource принимала веса как TArray<int32>. RarityWeight по
умолчанию 1 -- умножение "1 * 0.5" (доля клетки) усекалось бы до 0,
ингредиент с обычным весом стал бы невыбираемым на любом стыке двух
регионов, без единой ошибки в логе. Сигнатура переведена на
TArray<float> (тело функции не менялось -- она уже кастовала
BaseWeights[i] в float на каждой итерации, редундантные касты убраны).
GetRandomResourceForNiche (ниша сада, без пересечения биомов) -- просто
перевод int32->float перед вызовом, логика не меняется.

Одинаковый FName-кандидат из двух пересекающихся биомов -- сознательно
НЕ дедуплицируется, вероятностная масса складывается.

4 новых теста (IngredientRegistryBiomeWeightTest.cpp), тот же приём
CountPicks по многим прогонам, что уже в IngredientHarvestWindowTest.cpp
(переименован в CountBiomeWeightPicks -- та же ODR-коллизия unity-сборки,
что уже ловили раньше в сессии): пустой BiomeWeights совпадает со старым
однобиомным поведением; 0.5/0.5 с ДЕФОЛТНЫМ RarityWeight=1 у обоих --
прямой регресс-пин на находку выше (без правки B не выпал бы вообще);
перекос 0.9/0.1; общий кандидат в двух биомах получает суммарную
вероятностную массу.

186 -> 190 тестов. Полный прогон: 190/190, ноль регрессий.

+++ ProjectHerbalist/Source/ProjectHerbalist/Core/Subsystems/IngredientRegistrySubsystem.h
@@ -62,6 +62,15 @@ private:
     // — вынесена из GetRandomResourceForBiome, чтобы GetRandomResourceForNiche
     // не дублировал ту же формулу (Suitability по State + сезон/время/луна/
     // погода) над другим списком кандидатов.
-    FName PickWeightedResource(const TArray<FName>& Candidates, const TArray<int32>& BaseWeights,
+    //
+    // BaseWeights — TArray<float>, не int32 (правка 2026-08-31, PCG-биомы):
+    // GetRandomResourceForBiome домножает базовый вес карточки на долю
+    // клетки в биоме (Cell.BiomeWeights, может быть 0.5/0.33/...) до
+    // вызова этой функции — при исходном int32 умножение "1 (дефолтный
+    // RarityWeight) * 0.5" усекалось бы до 0, ингредиент с обычным весом
+    // становился бы невыбираемым на любом стыке двух регионов, без единой
+    // ошибки в логе. Сама функция ниже не меняется — она уже кастовала
+    // BaseWeights[i] в float на каждой итерации.
+    FName PickWeightedResource(const TArray<FName>& Candidates, const TArray<float>& BaseWeights,
         const FGridCell& Cell, const FHarvestContext& Context, FRandomStream& Rng) const;
 };
